Update GroupsClient tests to module-per-method
This splits the GroupsClient tests into a module per method.
